test(api): add unit tests for remaining mutation resolvers

+++ apps/api/src/resolvers/mutations/deleteThread.ts
@@ -3,6 +3,7 @@ import { UserInputError } from 'apollo-server-core';
 import { PrivateThread } from '@/db';
 import type { ObjectID } from '@uxc/types';
 import type { Resolver } from '../types';
+import { isValidObjectId } from 'mongoose';
 
 export const deleteThread: Resolver<ObjectID, { threadId: ObjectID }> = async (
 	_,
@@ -12,6 +13,14 @@ export const deleteThread: Resolver<ObjectID, { threadId: ObjectID }> = async (
 	if (!threadId) {
 		throw new UserInputError(ERROR_MESSAGES.E_NO_THREAD_ID);
 	}
+
+	if (!isValidObjectId(threadId)) {
+		throw new UserInputError(
+			`The provided threadId ${threadId} is not a valid ObjectID`
+		);
+	}
+
+	// @todo soft delete so we don't delete for both users
 	await PrivateThread.deleteOne({ _id: threadId });
 
 	return threadId;
